Add a function for trying next address
During non-blocking connect, isRead() tries the next addess, but
this closes the first socket, and opens another, causing an issue
with cyphesis as the socket closure cleans up the object.
Having the next attempt more explicit allows the app more control.
This has no effect on the existing behavior, as isRead() has not
yet been changed.